Fix Particle system

+++ KungFu/Assets/Scripts/Animation/BaseSmb.cs
@@ -9,7 +9,7 @@ public class BaseSmb : StateMachineBehaviour
         PlayerAnimController _animCtrl = animator.gameObject.GetComponent<PlayerAnimController>();
         if (_animCtrl != null)   // it is not a enemy
         {
-            animator.SetFloat("AttackID_i", -1);
+            animator.SetInteger("AttackID_i", -1);
             animator.SetInteger("DefenseID_i", -1);
         }
     }
